% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_coc.R
\name{county_coc}
\alias{county_coc}
\title{API Wrapper to call Colorado County Components of Change Data}
\usage{
county_coc(fips_list, year_list, group = "none")
}
\arguments{
\item{fips_list}{Numeric FIPS code(s) for the county (0 for the state) (no leading 0's)}

\item{year_list}{Numeric list of years between 1970 and 2050}

\item{group}{string taking values of opt1, opt2 see descriptiong what what each does}
}
\description{
This API wrapper returns a dataframe of county-level components of change data:
}
\details{
The functions takes county and year and can return grouped data.

Grouping options are as follows:
   opt1= Group by year
   opt2= Group by county

Variables include the estimate, popultion change, births, deaths, and net migration.

Data are estimates or forecasts depending on the most recent vintage, they are noted as such in the returned data.
}

