% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni_hist.R
\name{muni_hist}
\alias{muni_hist}
\title{API Wrapper to call Colorado historical municipal population estimates}
\usage{
muni_hist(fips_list = "", year_list, county = "",
  vars = "totalpopulation", totals = "yes")
}
\arguments{
\item{fips_list}{Numeric FIPS code(s) for the municipality(ies). Leave blank and specify county to get all places in a county Defaults to blank}

\item{year_list}{Numeric list of years between 1980 and 2015}

\item{county}{county FIPS number used to pull either parts of a muni or all munis in a county if fips_list is blank Defaults to Blank}

\item{vars}{list of variables to pull includes: totalpopulation, householdpopulation, groupquarterspopulation, householdsize totalhousingunits, occupiedhousingunits, and vacanthousingunits Defaults to totalpopulation}

\item{totals}{controls whether API returns totals or parts of municipalities Defaults to 'yes' to give totals not parts.}
}
\description{
This API wrapper returns a dataframe of municipal populations and related variables for the years
 and municipalities selected.
}
\details{
Note: Selecting all of the counties can be done by providing 300 as the fips_list
}

